requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autofaiss==2.15.6
backcall==0.2.0
black==23.3.0
blinker==1.6.2
braceexpand==0.1.7
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
-e git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1#egg=clip
clip-anytorch==2.5.2
cloudpickle==2.2.1
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
dataclasses==0.6
debugpy==1.6.7
decorator==5.1.1
-e git+https://github.com/facebookresearch/detectron2.git@af614bf8282ff14e159c45f37399e87b91a1ec41#egg=detectron2
diffusers==0.14.0
docker-pycreds==0.4.0
einops==0.3.0
embedding-reader==1.5.0
executing==1.2.0
ExifRead-nocycle==3.0.1
faiss-cpu==1.7.4
filelock==3.11.0
fire==0.4.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.53.0
h5py==3.8.0
huggingface-hub==0.14.1
hydra-core==1.3.2
idna==3.4
imageio==2.27.0
importlib-metadata==6.3.0
importlib-resources==5.12.0
iopath==0.1.9
ipykernel==6.22.0
ipython==8.9.0
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json-tricks==3.16.1
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mmcv==1.7.1
multidict==6.0.4
multilingual-clip==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numpy==1.23.2
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.20.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.1
pandas==1.5.3
parso==0.8.3
pathspec==0.11.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
portalocker==2.7.0
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pyDeprecate==0.3.2
Pygments==2.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.6.5
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
qudida==0.0.4
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
sentry-sdk==1.23.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
tabulate==0.9.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.3.21
timm==0.9.2
tokenizers==0.12.1
tomli==2.0.1
torch==1.10.1
torchaudio==0.10.1
torchmetrics==0.11.4
torchvision==0.11.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
types-pytz==2022.6.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
urwid==2.1.2
wandb==0.12.21
wcwidth==0.2.5
webdataset==0.1.103
widgetsnbextension==4.0.7
yacs==0.1.8
yapf==0.32.0
yarl==1.8.2
zipp==3.15.0
